requirements.txt
django>=5.0
pymysql>=1.1.0
cryptography>=41.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
